requirements.txt
opencv-contrib-python
numpy
Pillow
arabic-reshaper
python-bidi
flask
flask-cors
